+++ src/main/java/fields/node/properties/IntegerProperty.java
@@ -1,14 +1,12 @@
 package fields.node.properties;
 
 import fields.Identifier;
-import fields.node.Property;
 import org.jetbrains.annotations.NotNull;
 
-public class DoubleProperty extends MinMaxProperty<Double>{
-    public DoubleProperty(@NotNull Double min, @NotNull Double max) {
+public class IntegerProperty extends MinMaxProperty<Integer>{
+    public IntegerProperty(@NotNull Integer min, @NotNull Integer max) {
         super(min, max);
     }
-
     @Override
     protected @NotNull Identifier initType() {
         return new Identifier("brigadier","double");
